pull feature to its own class
